requirements.txt
requests==2.20.0
jsonpickle==0.7.1
cachecontrol==0.11.7
python-dateutil==2.5.3
responses~=0.22.0
urllib3~=1.24.3
setuptools~=65.5.1